nuevos endpoints de ventas gracias por choqlito

+++ models/sales.go
@@ -7,7 +7,7 @@ Structs de ventas
 type Sale struct {
 	ID       *string   `json:"id,omitempty"`
 	User     string    `json:"user"`
-	Total    int       `json:"total"`
+	Total    float64   `json:"total"`
 	Date     string    `json:"date"`
 	Products []Product `json:"products"`
 }
